Refactor: Remove unnecessary wrapper methods in Action classes
Removed unnecessary wrapper methods in Action classes like ShoulderActions, IntakeActions, ClawActions, WristActions, ElbowActions and LiftActions.
These wrapper methods were redundant as the inner static classes could be directly instantiated.
This change simplifies the code and makes it more readable.

Additionally, in the Auto class, an instance of Robot is now created to pass it to the LiftActions.LiftUp action.